Split the processing function to a separate module
We want to use unblob as a library, so we don't want to create the click.Command
object every time we import the process_file function.

It's also better for testing, as this import won't have side effects.
